bigint support

+++ isprime.c
@@ -1,6 +1,6 @@
 #include "prime.h"
-int isprime(int temp){
-	for(int i=2;i<temp;i++)
+int isprime(bigint temp){
+	for(bigint i=2;i<temp;i++)
 		if(!(temp % i))
 			return 0;
 	return 1;
