+++ prime.c
@@ -1,6 +1,6 @@
 #include "prime.h"
-void prime(int* in,int* out){
-	for(int i=2;i<(*in);i++)
+void prime(bigint* in,bigint* out){
+	for(bigint i=2;i<(*in);i++)
 		if(isprime(i))
 			if(!((*in) % i)){
 				*in=(*in) / i;
